Make style property Optional in TextyTextView
Making the style property Optional allows for TextyTextView to be initialized from a storyboard (`init(coder:)`). To make this work, this commit also introduces code that sets properties on the super class when a style does not exist. In theory, this allows for TextyTextView to be used exactly like UITextView up until the moment the style property is set. Although this is not the intended usage of TextyTextView, it is better than not doing anything with the new values when the style property does not exist.